move task-specific validator to plugin
